backend/pyproject.toml
[project]
name = "interview-practice-backend"
version = "0.1.0"
description = "AI-powered Interview Practice Application - Backend API"
authors = [
    {name = "Julian López", email = "jc.lopezs@uniandes.edu.co"}
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "openai>=1.109.1",
    "python-dotenv>=1.1.1",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "python-multipart>=0.0.6",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "anthropic>=0.69.0",
    "google-generativeai>=0.3.0",
    "langchain-community>=0.3.30",
    "wikipedia>=1.4.0",
    "duckduckgo-search>=8.1.1",
    "langgraph>=0.6.8",
    "faiss-cpu>=1.12.0",
    "chromadb>=1.1.1",
    "sentence-transformers>=5.1.1",
    "nltk>=3.9.2",
    "ruff>=0.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
